Parse VLAN and Bond configuration
With VLANs and Bonds (LAGs) we now have a multi-layer hierarchy of network ports.
VLANS and Bonds are classified as L2 adapters and compose a new layer in-between
SystemAdapters and PhysicalIO.
For every object that is either directly referenced or transitively used
by a SystemAdapter, parseSystemAdapterConfig() adds a separate NetworkPortConfig
entry into DevicePortConfig. Ports which are directly referenced by
SystemAdapters (i.e. without any higher-level ports above them), are called
"Endpoints". There is a new field "IsEndpoint" inside NetworkPortConfig
and NetworkPortStatus, that can be used to differentiate between an
endpoint interface and an intermediate (non-endpoint) port. This is usefull
because configuration for IP, DHCP, Proxy, Wireless, etc. is only relevant for
endpoint ports. Intermediate ports will only have labels and L2 configuration
provided in NetworkPortConfig.
Zedrouter should not do anything with intermediate ports.

This commits also clears up the semantics of logical labels and cross-layer
references.
The current implementation of the reference relations between system adapters
and physical IO is confusing. EVE code does not have a uniform way of looking up
for referenced objects. It mixes up adapter names with LowerLayerNames for
cross-layer references and it uses both logical and physical labels as targets
for those references.
The desired behaviour, which this commit enforces, is that controllers should
always have the upper layers reference lower layers by making the lowerLayerName
field of upper layers reference the logicalLabel field of lower layers.

Signed-off-by: Milan Lenco <milan@zededa.com>