Dependency graph for configuration items
This commit adds package "depgraph" under EVE/pillar implementing a dependency
graph [1]. It will be used in EVE to generically solve the problem of reconciliation
beetween the actual/current (configuration) state and the latest intended/desired
state. Everytime the desired state changes, a certain sequence of Create, Modify
and Delete operations has to be scheduled and executed to update the current
state. The order of operations often matters and it has to respect all dependencies
that exist between configuration items.
The plan is to first refactor NIM and use dependency graph for configuration
items that device networks are built with.
Based on how NIM refactoring goes, depgraph could be used for zedrouter next.

For more detailed reasoning and motivation behind depgraph, please refer to
the proposal presented on the LF Edge wiki [2].

[1]: https://en.wikipedia.org/wiki/Dependency_graph
[2]: TODO Proposal on wiki

Signed-off-by: Milan Lenco <milan@zededa.com>