Fix TLS config for non-controller diag GET tests
diag periodically performs GET requests towards the controller
to check the connectivity status and report it to the console.
If the controller connectivity is failing, diag will try to run GET
for google.com, both with HTTP and HTTPS, and the results are also
reported into the console. This is to separate between completely
broken connectivity and only controller-specific network issues
and make that known to the user.

However, it turned out that the TLS config used for google.com
carries SNI (Server Name Indication) of the controller, not "google.com".
And even though the server verification is intentionally disabled in this
case, SNI can still play a role in case there is a SNI-based transparent
HTTPS proxy in the way. An attempt to access google.com will get the same
treatment from the proxy as it would apply for the controller destination.
As a result the outcome of the test may be misleading.

Moreover, when diag runs tests towards google.com, it disables
server verification in the TLS config and this stays disabled even for
future controller tests, which is not desired. This commit fixes both bugs.

Signed-off-by: Milan Lenco <milan@zededa.com>