Support case with IP-addressed HTTP proxy and no DNS servers
If an explicit HTTPS proxy addressed by an IP address is used,
EVE does not need to perform any domain name resolution.
The resolution of the controller's domain name is performed by the proxy,
not by EVE. In this scenario EVE should therefore not treat missing DNS
server as an error.

Signed-off-by: Milan Lenco <milan@zededa.com>