Library for network tracing
This library, called nettrace, allows to trace (monitor and record a
summary of) network operations that happen behind the scenes during
a HTTP request processing as executed by http.Client. The collected
network traces are useful for troubleshooting of failed HTTP requests.

This commit only brings nettrace into the EVE repo under /libs, without
actually using it from the pillar yet.

For more information on how this is going to be used, please refer
to the proposal:
https://docs.google.com/document/d/1t-OxyjqjfwGbgyRWchaDfjRyUQ0r-LzqauHrppxB7Qk/edit?usp=sharing

Signed-off-by: Milan Lenco <milan@zededa.com>