Trace /config, /info and /ping zedcloud requests.
NIM and zedagent are enhanced to trace zedcloud requests using
the libs/nettrace package and publish them using the pkg/pillar/netdump
package.

Tracing introduces some overhead, therefore it is not enabled for every
request, but only for one request per hour before onboarding and one request
per day after onboarding (by default, but these intervals are configurable).

NIM traces /ping requests which are used to test controller
connectivity. Tracing is enabled only once per tracing-interval
(e.g. once per hour/day) and when the latest DPC is being tested.
When a trace is obtained, it is published inside a netdump archive
together with a packet capture (which is also enabled) into the netdump
topic nim-ok or nim-fail, depending on the outcome of the connectivity
test. When connectivity to the controller is not working, NIM also runs
traced requests towards google.com over HTTP and HTTPs and includes
traces from these requests as well. This can help to determine if
the issue is with the Internet access or with something specific to
the controller.

Similarly, zedagent runs tracing once per day for /config as well as for
/info requests (but only for ZInfoDevice). Tracing is enabled only when
the latest DPC is being used and reported by NIM as working.
If the basic connectivity is not working (/ping is failing), we will get
netdump from NIM, so there is no point in running tracing from zedagent
as well in such case.
Zedagent publishes these traces to netdump topics:
 * zedagent-info-ok/fail and
 * zedagent-config-ok/fail

All these netdumps end up being persisted in /persist/netdump as tgz
archives (see pkg/pillar/netdump package).
In order to not keep increasing the storage usage endlessly, each topic
is limited to a maximum of 10 netdumps (this number is also configurable).
Once this limit is reached for a given topic, a new netdump will kick
out the oldest one.

Signed-off-by: Milan Lenco <milan@zededa.com>