Separate and generalize object-to-number allocator
In pillar we now have 3 number allocators with persistence
built on top of pubsub publication:
- appnumallocator.go
- bridgenumallocator.go
- appnumonunet.go

Most of the code of these allocators is in zedrouter, but some underlying
functions are separated to pillar/uuidtonum and pillar/uuidpairtonum and
partially also used from other microservices.

The main logic of these allocators is the same, only target objects
for which the numbers are allocated differ.

The main problems with the current implementation are:
- hard to understand, not easy to read, type names like UUIDPairAndIfIdxToNum
  are quite clumsy
- same logic replicated 3 times
- difficult/impossible to re-use elsewhere (a concept of number
  allocator could be useful outside of zedrouter)
- some hard-coded limits (e.g. allocation limited to 256 numbers max)
- difficult/impossible to customize or extend
- most of the code is tightly-coupled with the zedrouter

In this commit we generalize all these allocators into one
implementation, separated from the zedrouter and put into pillar/objtonum
package.
The allocator itself was split into 3 types:
- object-to-number map: allows to store number assignments (e.g. using pubsub
  for persistence)
- number allocator: allows to allocate number from some integer set
  (could be a range of numbers <0-255> or something else)
- object-to-number allocator: allocates number for an object and stores
  the allocation. Uses an object-to-number map and a number allocator
  (of the caller choice).

This commit is first in a series of the zedrouter refactoring.

Signed-off-by: Milan Lenco <milan@zededa.com>