DPC Reconciler: patches for network routes and WLAN config.
This commit contains few improvements for the state reconciliation
of network routes. They are mostly aimed at avoiding unnecessary
Route Modify operations (which sometimes cause routes to be re-created
even when config has not actually changed but only some runtime route
flags make Route.Equal() to return false) and also to avoid some
false-negative error log messages (e.g. when trying to delete route
which was already automatically deleted by kernel).

Additionally, there is a change in the description outputted for WLAN
config in the dependency graph to NOT include the hashed password.
This is stored in /run/nim-current-state.dot and /run/nim-intended-state.dot.
Even though this filesystem is only in memory (just like
/run/wlan/wpa_supplicant.conf), it is safer to avoid spreading this secret
information into multiple files, because it increases the chance of
the password being accidentally leaked (e.g. by someone debugging
a connectivity issue and publishing this file to recipients who are
not intended to know the password, even if it is already hashed - enough
to connect to the WiFi network).

Signed-off-by: Milan Lenco <milan@zededa.com>