Remove old IPs from the bridge before assigning new ones
This is just a silly hack to avoid an issue in dnsmasq - the "listen-address"
attribute that we put into the config is effectively ignored. DNS server will
listen on all IPs assigned to the bridge interface anyway. The problem is that
as Reconciler moves from one intended state to another with changed NI subnet,
a temporary intermediate state may be reached where bridge has both the old
and the new IPs assigned and dnsmasq will start and continue listening on all IPs
even after the old IPs are removed - this makes DNS port on old IPs occupied
and prevent other NIs to use them.
This hack prevents intermediate reconciliation states with bridges having old
and new IPs assigned at the same time - old IPs are forced to be removed before
new ones are added.

Signed-off-by: Milan Lenco <milan@zededa.com>