wwan: allow to disable periodic querying of visible providers
The operation to retrieve the set of visible cellular providers is quite
expensive and takes around 2 minutes to complete, during which the modem
is not manageable. After recent improvement in the wwan microservice
we now separately publish info about the currently used provider, which
is not so expensive to obtain. Therefore, there is little incentive left
to query all visible providers now. For the user it makes sense to
enable it only temporarily to for example troubleshoot modem failing to
register in the network.
In this commit we make the operation to query visible providers
configurable and disabled by default.

Signed-off-by: Milan Lenco <milan@zededa.com>