Implement features of the new cellular APIs
This is the second part of the commit series implementing the new
cellular APIs in EVE. The first part was mostly about extending zedagent
microservice to support parsing of the new protobuf messages and extending
structures used for delivering config and status between NIM and wwan
microservice. But wwan and for the most part also NIM still worked only
in the backward-compatible mode.
In this commit, we make substantial enhancements to the wwan microservice
and make few changes to NIM to implement the new cellular APIs in their
full scope.

Part II changelog:
* implemented username/password authentication for cellular networks
* password is transferred from NIM to wwan encrypted just to avoid
  accidental exposure when debug info is collected and sent outside
  (otherwise not necessary because the underlying storage is in memory)
* added logging with configurable verbose level to wwan microservice
* upgraded QMI and MBIM CLI tools to bring in some later added commands
  needed to implement the new cellular APIs. Note that some of the patches
  we were applying to the previously used versions are already included
  in the newer versions and can be therefore removed from the EVE repo
* improved robustness of the wwan microservice - decreased reaction time
  and improved ability to recover from lost/broken connectivity
* finalized support for multiple cellular modems (just few changes were
  needed, most of the code was already there)
* more cellular status data are now published, e.g.: time when the
  connection was established, info about currently used provided,
  used Radio Access Technology (UMTS / LTE / ...), modem manufacturer,
  etc.
* added support for modems with multiple SIMs in the DSSS mode
  (single standby)

Signed-off-by: Milan Lenco <milan@zededa.com>