NI Reconciler: allow to specify VIF host ifname from upper layers
With kubernetes networking, the virtual network interface (VIF) is
created between an application and the host by a CNI plugin before
zedrouter is called to establish app connectivity. This means that by
the time ConnectApp of NI Reconciler is called, the VIF name is already
given and should not be generated. In this commit, we enable to specify
VIF name as part of arguments for ConnectApp.
When empty (i.e. not kubernetes deployment), NI Reconciler will
generate interface name as before.

Signed-off-by: Milan Lenco <milan@zededa.com>