Parse, validate and report MTU config
This commits adds MTU attribute to various structs with network
config or status in pillar/types.
Parsing of the MTU config performed by zedagent is also included.
The actual implementation for MTU will be done in subsequent commits.
Changes in pillar/types had to be done first and separately so that they
can be imported from pkg/wwan without using temporary replace directive.

Signed-off-by: Milan Lenco <milan@zededa.com>