Force IsPCIBack re-check when network interface appears in the kernel
We have some rare instances where hyper.PCIRelease() called to return
network port from PCIBack takes longer to take effect. But we immediately
check the network interface presence and leave IsPCIBack as true if
the interface is not found. IsPCIBack may remain in this incorrect
state for a fairly long time. This is a problem because NIM does not
configure ports which domainmgr claims to be in the PCIBack.

As a workaroung, let's add interface index into DNS. When the interface
finally appears, this will get non-zero value, triggering DNS update
and as a result domainmgr will update IsPCIBack flag in the
AssignableAdapters struct.

Signed-off-by: Milan Lenco <milan@zededa.com>