Do not remove lastresort config until EVE gets usable DPC
By default the lastresort network config (try all ethernet interfaces
with DHCP) is disabled and used only if there is no other network config
available. This is usually the case on the first device boot. However,
if device received DPC but without a single management port with a valid
config, then lastresort should still keep being used and not removed.

Signed-off-by: Milan Lenco <milan@zededa.com>